Unify two classes called CleaningRandomAccessFileTest
